refactor: move methods from views to utils
